+++ lib/influxdb2/client/flux_csv_parser.rb
@@ -22,6 +22,7 @@ require 'base64'
 
 module InfluxDB2
   # This class represents Flux query error
+  #
   class FluxQueryError < StandardError
     def initialize(message, reference)
       super(message)
@@ -32,6 +33,7 @@ module InfluxDB2
   end
 
   # This class represents Flux query error
+  #
   class FluxCsvParserError < StandardError
     def initialize(message)
       super(message)
@@ -39,20 +41,28 @@ module InfluxDB2
   end
 
   # This class us used to construct FluxResult from CSV.
+  #
   class FluxCsvParser
-    def initialize
+    include Enumerable
+    def initialize(response, stream: false)
+      @response = response
+      @stream = stream
       @tables = {}
 
       @table_index = 0
       @start_new_table = false
       @table = nil
       @parsing_state_error = false
+
+      @closed = false
     end
 
-    attr_reader :tables
+    attr_reader :tables, :closed
 
-    def parse(response)
-      CSV.parse(response) do |csv|
+    def parse
+      @csv_file = CSV.new(@response.instance_of?(Net::HTTPOK) ? @response.body : @response)
+
+      while (csv = @csv_file.shift)
         # Response has HTTP status ok, but response is error.
         next if csv.empty?
 
@@ -68,10 +78,24 @@ module InfluxDB2
           raise FluxQueryError.new(error, reference_value.nil? || reference_value.empty? ? 0 : reference_value.to_i)
         end
 
-        _parse_line(csv)
+        result = _parse_line(csv)
+
+        yield result if @stream && result.instance_of?(InfluxDB2::FluxRecord)
       end
 
-      @tables
+      self
+    end
+
+    def each
+      return enum_for(:each) unless block_given?
+
+      parse do |record|
+        yield record
+      end
+
+      self
+    ensure
+      _close_connection
     end
 
     private
@@ -84,7 +108,9 @@ module InfluxDB2
         # Return already parsed DataFrame
         @start_new_table = true
         @table = InfluxDB2::FluxTable.new
-        @tables[@table_index] = @table
+
+        @tables[@table_index] = @table unless @stream
+
         @table_index += 1
       elsif @table.nil?
         raise FluxCsvParserError, 'Unable to parse CSV response. FluxTable definition was not found.'
@@ -157,13 +183,17 @@ module InfluxDB2
           @table.columns.push(column)
         end
 
-        @tables[@table_index] = @table
+        @tables[@table_index] = @table unless @stream
         @table_index += 1
       end
 
       flux_record = _parse_record(@table_index - 1, @table, csv)
 
-      @tables[@table_index - 1].records.push(flux_record)
+      if @stream
+        flux_record
+      else
+        @tables[@table_index - 1].records.push(flux_record)
+      end
     end
 
     def _parse_record(table_index, table, csv)
@@ -205,6 +235,12 @@ module InfluxDB2
       else
         str_val
       end
+    end
+
+    def _close_connection
+      # Close CSV Parser
+      @csv_file.close
+      @closed = true
     end
   end
 end