+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 ## 1.1.0 [unreleased]
+
+### Features
+1. [#14](https://github.com/influxdata/influxdb-client-ruby/issues/14): Added QueryApi
  
 ## 1.0.0.beta [2020-01-17]